Added Banner and Map elements
