+++ app/javascript/packs/components/Header.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Header extends React.Component {
+  render() {
+    return (
+      <div>
+        <img className="page-header" src="TenBisBig.png" />
+      </div>
+    );
+  }
+}
+
+export default Header;
